requirements.txt
Django==1.5.1
PyVCF==0.6.3
argparse==1.2.1
distribute==0.6.36
psycopg2==2.4.6
wsgiref==0.1.2
